# src/main/resources/api/api-docs.yaml
openapi: 3.0.3
info:
  title: PlayMatch API
  description: API for PlayMatch service
  version: 1.0.0
  contact:
    name: PlayMatch Team
    url: https://playmatch.com
    email: support@playmatch.com
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.playmatch.com
    description: Production server
...
paths:
...
  /v1/auth/register:
    post:
      tags:
        - Auth
      summary: Create user by email & password
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: User created
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/login:
    post:
      tags:
        - Auth
      summary: Login with email & password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login success
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "423": { $ref: "#/components/responses/Locked" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Start password reset flow
      description: Always returns 202 to avoid user enumeration.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        "202":
          description: Accepted (email sent if account exists)
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /v1/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Complete password reset using token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "400": { $ref: "#/components/responses/BadRequest" }
        "410": { $ref: "#/components/responses/Gone" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: |
        Generates new access and refresh tokens using a valid refresh token.
        The refresh token can be provided either in the request body or via httpOnly cookie.
        Cookie-based approach is preferred for better security against XSS attacks.
      operationId: refreshToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: JWT refresh token (optional if using cookie)
                  example: "eyJhbGciOiJIUzI1NiJ9..."
      responses:
        "200":
          description: Token refreshed successfully
          headers:
            X-Request-Id: { schema: { type: string } }
            Set-Cookie:
              description: httpOnly cookie with new refresh token (if cookie support enabled)
              schema:
                type: string
                example: "refreshToken=eyJhbGciOiJIUzI1NiJ9...; HttpOnly; Secure; Path=/v1/auth; Max-Age=604800; SameSite=Strict"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410": { $ref: "#/components/responses/Gone" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # Player Profile paths
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, gender, email, password, reEnterPassword]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128 }
        reEnterPassword: { type: string, minLength: 8, maxLength: 128 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, example: 3600 }
        refreshToken: { type: string, nullable: true }
        user: { $ref: "#/components/schemas/AuthUser" }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, password, reEnterPassword]
      properties:
        token: { type: string }
        password: { type: string, minLength: 8, maxLength: 128 }
        reEnterPassword: { type: string, minLength: 8, maxLength: 128 }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
    AuthUser:
      type: object
      properties:
        id: { type: string, format: long }
        name: { type: string }
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://api.example.com/errors/validation" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
  responses:
    BadRequest:
      description: Validation or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Conflict:
      description: Email already in use
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Locked:
      description: Account temporarily locked due to too many attempts
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Gone:
      description: Token expired or already used
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }